Reorganize package, add greeting message to GUI.
