+++ src/main/java/duke/logic/task/ToDo.java
@@ -1,4 +1,4 @@
-package duke.task;
+package duke.logic.task;
 
 /**
  * Represents a to-do task.
