+++ src/test/java/duke/task/TodoTest.java
@@ -1,5 +1,6 @@
 package duke.task;
 
+import duke.logic.task.ToDo;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
